Parse constants from cmd flags directly

+++ api/server.go
@@ -15,7 +15,7 @@ import (
 )
 
 // New returns a new instance of the server. To execute it, the method `ListenAndServe` must be called.
-func New(port uint16, localServer bool, devMode bool, logRequests bool, db *gorm.DB, manualUpdate chan *models.Podcast) *http.Server {
+func New(port uint, localServer bool, devMode bool, logRequests bool, db *gorm.DB, manualUpdate chan *models.Podcast) *http.Server {
 	if db == nil {
 		log.Panic("'podcastsDB' is nil")
 	}
